refactor(auth): promote the @Public marker to shared
Core operational routes (health probes, the not-found fallback) marked
themselves public by importing a decorator from the auth feature, pointing
global infrastructure at a business domain.

The marker is pure route metadata with no authentication logic, and
SessionGuard still owns its interpretation, so it belongs in shared/.
