--- src/features/auth/decorators/public.decorator.ts
@@ -1,5 +1,0 @@
-import { SetMetadata } from '@nestjs/common';
-
-export const IS_PUBLIC_KEY = 'isPublic';
-
-export const Public = () => SetMetadata(IS_PUBLIC_KEY, true);
